feat: reservation 테이블 접근하는 메서드 위치 변경
